Reverse merge beta
