+++ src/components/platformSpecific/googleSheetsPage.js
@@ -1,6 +1,8 @@
 function renderGoogleSheetsPage({ manifest, userSettings }) {
     const existingGoogleSheetsName = userSettings?.googleSheetsName?.value ?? "";
     const existingGoogleSheetsUrl = userSettings?.googleSheetsUrl?.value ?? "";
+    const isManaged = userSettings?.googleSheetsName?.customizable === false && userSettings?.googleSheetsUrl?.customizable === false;
+    
     const page = {
         id: 'googleSheetsPage',
         title: "Google Sheets Config",
@@ -10,14 +12,16 @@ function renderGoogleSheetsPage({ manifest, userSettings }) {
             properties: {
                 warning: {
                     type: "string",
-                    description: "Only one sheet can be used at the same time."
+                    description: isManaged 
+                        ? "This Google Sheet is managed by your administrator and cannot be changed."
+                        : "Only one sheet can be used at the same time."
                 }
             }
         },
         uiSchema: {
             warning: {
                 "ui:field": "admonition",
-                "ui:severity": "warning",  // "warning", "info", "error", "success"
+                "ui:severity": isManaged ? "info" : "warning",  // "warning", "info", "error", "success"
             }
         },
         formData: {
@@ -25,10 +29,7 @@ function renderGoogleSheetsPage({ manifest, userSettings }) {
         }
     }
     if (existingGoogleSheetsName) {
-        let removeSheetButtonDisabled = false;
-        if (userSettings?.googleSheetsName?.customizable === false && userSettings?.googleSheetsUrl?.customizable === false) {
-            removeSheetButtonDisabled = true;
-        }
+        let removeSheetButtonDisabled = isManaged;
         page.schema.properties.sheetInfoButton = {
             type: "string",
             title: `Sheet name: ${existingGoogleSheetsName}`
@@ -52,32 +53,35 @@ function renderGoogleSheetsPage({ manifest, userSettings }) {
         page.formData.sheetUrl = existingGoogleSheetsUrl;
     }
     else {
-        page.schema.properties.newSheetName = {
-            type: "string",
-            title: "New sheet name"
-        }
-        page.schema.required = ["newSheetName"];
-        page.schema.properties.newSheetButton = {
-            type: "string",
-            title: "Create new sheet"
-        }
-        page.uiSchema.newSheetName = {
-            "ui:placeholder": 'Enter name...',
-        }
-        page.uiSchema.newSheetButton = {
-            "ui:field": "button",
-            "ui:variant": "contained", // "text", "outlined", "contained", "plain"
-            "ui:fullWidth": true,
-            "ui:disabled": true
-        }
-        page.schema.properties.selectExistingSheetButton = {
-            type: "string",
-            title: "Select existing sheet"
-        }
-        page.uiSchema.selectExistingSheetButton = {
-            "ui:field": "button",
-            "ui:variant": "contained", // "text", "outlined", "contained", "plain"
-            "ui:fullWidth": true
+        // Only show create/select options if not managed by admin
+        if (!isManaged) {
+            page.schema.properties.newSheetName = {
+                type: "string",
+                title: "New sheet name"
+            }
+            page.schema.required = ["newSheetName"];
+            page.schema.properties.newSheetButton = {
+                type: "string",
+                title: "Create new sheet"
+            }
+            page.uiSchema.newSheetName = {
+                "ui:placeholder": 'Enter name...',
+            }
+            page.uiSchema.newSheetButton = {
+                "ui:field": "button",
+                "ui:variant": "contained", // "text", "outlined", "contained", "plain"
+                "ui:fullWidth": true,
+                "ui:disabled": true
+            }
+            page.schema.properties.selectExistingSheetButton = {
+                type: "string",
+                title: "Select existing sheet"
+            }
+            page.uiSchema.selectExistingSheetButton = {
+                "ui:field": "button",
+                "ui:variant": "contained", // "text", "outlined", "contained", "plain"
+                "ui:fullWidth": true
+            }
         }
     }
     return page;
