+++ src/eventHandlers/rc-post-message-request/custom-button-click/googleSheetsConfig.js
@@ -3,7 +3,11 @@ import googleSheetsPage from '../../../components/platformSpecific/googleSheetsP
 
 async function onEvent({ data, manifest, platformInfo, platformName, platform }) {
     window.postMessage({ type: 'rc-log-modal-loading-on' }, '*');
-    const { userSettings } = await userCore.refreshUserSettings({});
+    const userSettings = await userCore.refreshUserSettings({});
+    if (!userSettings) {
+        window.postMessage({ type: 'rc-log-modal-loading-off' }, '*');
+        return;
+    }
     document.querySelector("#rc-widget-adapter-frame").contentWindow.postMessage({
         type: 'rc-adapter-register-customized-page',
         page: googleSheetsPage.renderGoogleSheetsPage({ manifest, userSettings })
